# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MatterCreate": {
        "properties": {
          "ai_processing_policy": {
            "default": "standard",
            "description": "AI processing policy",
            "title": "Ai Processing Policy",
            "type": "string"
          },
          "case_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Type of case",
            "title": "Case Type"
          },
          "client_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Client name",
            "title": "Client Name"
          },
          "confidentiality_level": {
            "default": "standard",
            "description": "Confidentiality level",
            "title": "Confidentiality Level",
            "type": "string"
          },
          "internal_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Internal reference ID",
            "title": "Internal Reference"
          },
          "jurisdiction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The jurisdiction of the matter",
            "title": "Jurisdiction"
          },
          "title": {
            "description": "The title of the matter",
            "maxLength": 255,
            "minLength": 3,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "MatterCreate",
        "type": "object"
      },
...
      "MatterResponse": {
        "properties": {
          "access_scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Scope"
          },
          "ai_processing_policy": {
            "title": "Ai Processing Policy",
            "type": "string"
          },
          "case_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Type"
          },
          "client_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Name"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "confidentiality_level": {
            "title": "Confidentiality Level",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "internal_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internal Reference"
          },
          "jurisdiction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdiction"
          },
          "opened_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opened At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "confidentiality_level",
          "ai_processing_policy"
        ],
        "title": "MatterResponse",
        "type": "object"
      },
...
      "UploadIntentRequest": {
        "properties": {
          "filename": {
            "maxLength": 255,
            "title": "Filename",
            "type": "string"
          },
          "matter_id": {
            "title": "Matter Id",
            "type": "string"
          },
          "mime_type": {
            "maxLength": 100,
            "title": "Mime Type",
            "type": "string"
          },
          "size_bytes": {
            "description": "File size in bytes",
            "exclusiveMinimum": 0.0,
            "maximum": 104857600.0,
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "matter_id",
          "filename",
          "mime_type",
          "size_bytes"
        ],
        "title": "UploadIntentRequest",
        "type": "object"
      },
      "UploadIntentResponse": {
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "presigned_url": {
            "title": "Presigned Url",
            "type": "string"
          },
          "revision_id": {
            "title": "Revision Id",
            "type": "string"
          },
          "storage_key": {
            "title": "Storage Key",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "revision_id",
          "presigned_url",
          "storage_key"
        ],
        "title": "UploadIntentResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "MESA Law API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents/upload-intent": {
      "post": {
        "operationId": "createUploadIntent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadIntentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadIntentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Upload Intent"
      }
    },
...
    "/api/v1/matters": {
      "get": {
        "operationId": "listMatters",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MatterResponse"
                  },
                  "title": "Response Listmatters",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Matters"
      },
      "post": {
        "operationId": "createMatter",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatterCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Matter"
      }
    },
...
  }
}